+++ R/sanity_model.R
@@ -98,6 +98,7 @@ sanity_model_supported_class <- function(model) {
         "speedglm",
         "speedlm",
         "stanreg",
+        "survreg",
         "svyolr",
         c("tobit", "survreg"),
         "tobit1",