+++ R/type_dictionary.R
@@ -123,6 +123,9 @@ speedglm,response
 speedglm,link
 stanreg,response
 stanreg,link
+survreg,response
+survreg,link
+survreg,quantile
 svyglm,response
 svyglm,link
 svyolr,probs